requirements.txt
opencv-python
ultralytics
numpy
Pillow
shapely
mysql-connector-python
